lib.rs and unit tests because we're grown ups
